requirements.txt
swibots
python-dotenv
aiohttp
pymongo
telegraph